+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 
 ### BREAKING CHANGES
 
+
 - `\Wizaplace\SDK\Translation\TranslationService::pushXliffCatalog` now uses a `$password` parameter instead of the apiClient authentication
 - Removed `\Wizaplace\SDK\Catalog\Product::getUrl`
 - Removed `\Wizaplace\SDK\Catalog\ProductSummary::getUrl`
 - Removed `\Wizaplace\SDK\Basket\BasketItem::getProductUrl`
+- Replaced `\Wizaplace\SDK\Catalog\Company` by `\Wizaplace\SDK\Catalog\CompanySummary`
 
 ### New features